backend/requirements.txt
coverage
Django
django-cors-headers
django-filter
djangorestframework
pillow
psycopg2-binary
python-dotenv